refactor: [UF-358] Indicator 컴포넌트 variants 통합 및 타입 에러 수정
