+++ src/shared/ui/Indicator/Indicator.types.ts
@@ -1,15 +1,14 @@
-import { VariantProps } from 'class-variance-authority';
+import { ComponentProps } from 'react';
 
-import { basicDotVariants } from './indicatorVariants';
-
-export interface BasicDotIndicatorProps
-  extends React.HTMLAttributes<HTMLDivElement>,
-    VariantProps<typeof basicDotVariants> {
-  step: number; // 현재 단계 (1부터 시작)
-  totalSteps: number;
+export type BasicDotIndicatorProps = ComponentProps<'div'> & {
+  step?: number;
+  totalSteps?: number;
   size?: 'sm' | 'md' | 'lg';
   orientation?: 'horizontal' | 'vertical';
   spacing?: 'tight' | 'normal' | 'wide';
   showHover?: boolean;
   onStepClick?: (step: number) => void;
-}
+};
+
+export type BasicDotSize = 'sm' | 'md' | 'lg';
+export type BasicDotState = 'pending' | 'active' | 'completed';
