+++ src/shared/ui/Indicator/index.ts
@@ -1,3 +1,3 @@
 export { Indicator } from './Indicator';
-export type { BasicDotIndicatorProps } from './Indicator.types';
-export { indicatorVariants, basicDotVariants } from './indicatorVariants';
+export { indicatorVariants, basicDotVariants } from './IndicatorVariants';
+export type { BasicDotIndicatorProps, BasicDotSize, BasicDotState } from './Indicator.types';
